* updated port apis for SAMD21J18A * correction in analogueout api * test support for analogueout in mbed SDK

+++ libraries/mbed/targets/hal/TARGET_Atmel/TARGET_SAM0/TARGET_SAMD21J18A/analogout_api.c
@@ -103,5 +103,5 @@ uint16_t analogout_read_u16(dac_t *obj)
     uint32_t data_val = 0;
     Dac *const dac_module = (uint32_t)obj->dac;
     data_val = dac_module->DATA.reg;
-    return (uint16_t)((data_val /MAX_VAL_10BIT) * 0xFFFF);   /*Normalization to the value 0xFFFF*/
+    return (uint16_t)((data_val / (float)MAX_VAL_10BIT) * 0xFFFF);   /*Normalization to the value 0xFFFF*/
 }
